requirements.txt
anyio==3.7.1
APScheduler==3.10.1
attrs==25.3.0
blinker==1.6.2
certifi==2025.6.15
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
py-dotenv==0.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
PyVirtualDisplay==3.0
requests==2.31.0
selenium==4.34.2
setuptools==80.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.5.0
webdriver-manager==3.9.1
websocket-client==1.8.0
Werkzeug==2.3.6
wsproto==1.2.0
